Mobile No Update

+++ src/main/java/com/ing/service/CustomerService.java
@@ -1,10 +1,12 @@
 package com.ing.service;
 
+import java.math.BigInteger;
+
 import com.ing.dto.CustomerDto;
 
 public interface CustomerService {
 	
 	public String createAccount(CustomerDto customerDto);
 	
-	public String updateUser(String email,Long mobile,Long customerId);
+	public String updateUser(String email,BigInteger mobile,Long customerId);
 }
